+++ src/unit/unit_manager.cpp
@@ -114,7 +114,7 @@ void CUnitManager::ReleaseUnit(CUnit &unit)
 		unit.UnitManagerData.unitSlot = -1;
 		units.pop_back();
 	}
-	Assert(unit.PlayerSlot == -1);
+	Assert(unit.PlayerSlot == static_cast<size_t>(-1));
 	releasedUnits.push_back(&unit);
 	unit.ReleaseCycle = GameCycle + 500; // can be reused after this time
 	//Refs = GameCycle + (NetworkMaxLag << 1); // could be reuse after this time
